requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-chess==1.999
websockets==12.0
jinja2==3.1.3
python-multipart==0.0.9 